refactor(tui): split modal key handling out of multi_pane::keys
The agents overlay, rename prompt, share panel, and delete-tab
confirmation each take the keyboard away from the panes; their handlers
and tests now sit together in tui::multi_pane::modal.
